+++ core/src/main/java/io/snabble/sdk/SnabbleInitializer.kt
@@ -25,13 +25,6 @@ class SnabbleInitializer : Initializer<Snabble> {
             }
         }
 
-        fun Properties.getBoolean(key: String, default: Boolean) =
-            getProperty(key).toBooleanStrictOrNull() ?: default
-        fun Properties.getLong(key: String, default: Long) =
-            getProperty(key).toLongOrNull() ?: default
-        fun Properties.getFloat(key: String, default: Float) =
-            getProperty(key).toFloatOrNull() ?: default
-
         // load properties created by the gradle plugin
         val path = propertiesFiles[UserPreferences(context).environment]
         if (path != null) {
